nhl-game-predictor-backend/pyproject.toml
[tool.poetry]
name = "nhl-game-predictor-backend"
version = "0.1.0"
description = ""
authors = ["Evan Boyd <evan.boyd@mail.mcgill.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.2"
psycopg2 = "^2.9.10"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
httpx = "^0.27.2"
datetime = "^5.5"
scikit-learn = "^1.5.2"
pandas = "^2.2.3"
djangorestframework = "^3.15.2"
lime = "^0.2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
